+++ github/github-check-update.php
@@ -10,7 +10,7 @@ defined( 'ABSPATH' ) || die();
 
 require_once __DIR__ . '/plugin-update-checker/plugin-update-checker.php';
 use YahnisElsts\PluginUpdateChecker\v5\PucFactory;
-use YahnisElsts\PluginUpdateChecker\v5p6\Vcs\GitHubApi;
+use YahnisElsts\PluginUpdateChecker\v5p7\Vcs\GitHubApi;
 
 $ghupuc_token = ghupuc_query_token_needed();